+++ apps/store/src/features/memberArea/components/Header.tsx
@@ -1,0 +1,32 @@
+import styled from '@emotion/styled'
+import { t } from 'i18next'
+import { HedvigLogo, mq, theme } from 'ui'
+import { LogoLink, LogoWrapper } from '@/components/Header/Header'
+import { MENU_BAR_HEIGHT_DESKTOP, MENU_BAR_HEIGHT_MOBILE } from '@/components/Header/HeaderStyles'
+import { PageLink } from '@/utils/PageLink'
+
+export const Header = () => (
+  <HeaderWrapper>
+    <LogoWrapper>
+      <LogoLink href={PageLink.home()} aria-label={t('HOME_PAGE_LINK_LABEL')}>
+        <HedvigLogo />
+      </LogoLink>
+    </LogoWrapper>
+  </HeaderWrapper>
+)
+
+const HeaderWrapper = styled.header({
+  width: '100%',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+
+  top: 0,
+
+  height: MENU_BAR_HEIGHT_MOBILE,
+  paddingInline: theme.space.md,
+  [mq.lg]: {
+    height: MENU_BAR_HEIGHT_DESKTOP,
+    paddingInline: theme.space.xl,
+  },
+})
